Выносит компонент Cards
